+++ packages/orq-rc/src/models/operations/filelist.ts
@@ -125,7 +125,7 @@ export const FileListData$inboundSchema: z.ZodType<
   file_name: z.string(),
   workspace_id: z.string(),
   created: z.string().datetime({ offset: true }).default(
-    "2026-02-23T14:47:24.606Z",
+    "2026-02-23T15:59:55.483Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {